Establish docker-compose.yml for local development, tweak Dockerfile & entrypoint for local use
This establishes a docker-compose.yml file with a db service running
postgres, along with related changes to config/database.yml.

Because we need the db container to not only be running, but also
actually open on TCP port 5432, we add a bin/wait-for-tcp which is
invoked in a new bin/docker-dev-entrypoint (specified in the
docker-compose.yml).

The Rails default Dockerfile is intended for production use, and
required some tweaks to work in a local context:

1. Instead of hardcoding RAILS_ENV to production and BUNDLE_WITHOUT to
   development:test, we establish these as build args (defaulting to the
   production values) which can be overridden in docker-compose.yml for
   a local context.
2. Before rails is run, there won't be a Gemfile.lock or yarn.lock file,
   which causes the default COPY commands to fail. By adding * at the
   end, we ignore any copy issues.
3. Asset precompiling is done by default, but this is not necessary in a
   local context. This changes that RUN line to skip compilation unless
   we are RAILS_ENV=production.